Initial prototypes for MonarchyPOS uploaded
These are python prototypes for the MonarchyPOS I am writing in Java. They will be implemented eventually once all the bugs are worked out and the code is generally nicer. Cinema.py is the main prototyping program where the other prototypes eventually merge with and lists.py is currently just for defining certain functions and using lists for a better selection menu. Currently the main program uses switch/match statements, lists.py is prototyping a better way of handling it.